Advice

Description

Module providing helpful tips about minimizing the complexity of the chosen data structure

Documentation

data Advice

Type for advice data

Constructors

Advice

Fields

advisedDS :: Structure

Structure to be advised

reducedOperations :: [OperationName]

Operations on which the structure is better than the recommended one

operations :: [OperationName]

Original operations (we can get recommendations from them)

notWorse :: Structure -> Structure -> [OperationName] -> Bool

Checks if structure s1 is not worse than structure s2 on operations opns

better :: Structure -> Structure -> [OperationName] -> Bool

Checks if the structure s1 is better than the structure s2 on operations opns

betterThanEach :: Structure -> [Structure] -> [OperationName] -> Bool

Checks if the structure s1 is better than each of the structures in ss on operations opns

filterAdviceForRecommended :: [Structure] -> [Advice] -> [Advice]

Removes already recommended data structures recs from the advised structures

adviceDS' :: Integer -> [OperationName] -> [Advice]

Returns advice data for operations opns and at most n operations to forget about

adviceDS :: [OperationName] -> [Advice]

Returns advice data for operations opns

printAdvice' :: (String -> IO ()) -> Integer -> [OperationName] -> IO ()

Pretty prints effects of adviceDS'

printAdviceStructure :: (String -> IO ()) -> Advice -> IO ()

Prints one Advice element

printAdvice :: (String -> IO ()) -> [OperationName] -> IO ()

Pretty prints effects of adviceDS