Analyzer

Documentation

data DSUse

Data structure use case

Constructors

DSU

Fields

getDSUName :: OperationName

Operation used

isHeavilyUsed :: Bool

Is it heavily used

isUserDependent :: Bool

Is it dependent on some external input (user, network, random, signals, etc.)

data DSInfo

Data structure for analysis info for one data-structure (possibly in many forms of different variables in functions)

Constructors

DSI

Fields

getDSINames :: [(FunctionName, VariableName)]

Variables, used in functions, holding the analyzed data structure

getDSIDSU :: [DSUse]

DSUses of the data structure

data DSFun t

Data structure for analysis info of a function definition

Constructors

DSF

Fields

getDSFFun :: FunctionDeclaration t

Analyzed function declaration

getDSFCalls :: [FunctionCall]

FunctionCalls from the analyzed function

getDSFDSI :: [DSInfo]

DSInfo about the variables inside of the function, at this stage are not yet ready to obtain the information

data FunctionDeclaration t

Type for function definitions

Constructors

FunDecl

Fields

getFunName :: FunctionName

Name of the function

getFunType :: t

Return type of the function

getFunArgs :: [(VariableName, t)]

Names and types of the function arguments

type FunctionCall = (FunctionName, [Maybe VariableName])

Function call - name of the function, relevant arguments

type TermAnalyzer a = State TermAnalyzerState a

State monad with TermAnalyzerState

type Output = [(VariableName, DSUse)]

Term analyzer basic output, variables and DSUses

data TermAnalyzerState

State of the analyzer

Constructors

AS

Fields

getStateCalls :: [FunctionCall]

FunctionCalls gathered through the analysis

startingFunction :: FunctionName

Name of the starting function

stupidMerge :: [DSInfo] -> [DSInfo]

Stupid merging of dsis --TODO remove this function, rewrite analyzeFunctions correctly

analyzeFunctions :: [DSFun t] -> [DSInfo]

Merges the simple DSInfos based on function calls from the functions

analyzeDSF :: [DSFun t] -> [DSInfo]

Analyze wrapper with stupidMerge

lookupDSF :: [DSFun t] -> FunctionName -> DSFun t

Lookup DSFun by FunctionName

lookupDSI :: [DSInfo] -> VariableName -> FunctionName -> DSInfo

Lookup DSInfo by FunctionName and VariableName

bindFuncall :: [FunctionDeclaration t] -> FunctionName -> [Maybe VariableName] -> [(VariableName, VariableName)]

Returns pairs of local variables bound to variables in a function that is called

generateDSI :: FunctionName -> Output -> [DSInfo]

Generates simple DSInfos without the info from function calls