CAnalyzer

Documentation

data DSInfo

Data structure for analysis info for one data-structure (possibly in many forms of different variables in functions)

Constructors

DSI

Fields

getDSINames :: [(FunctionName, VariableName)]

Variables, used in functions, holding the analyzed data structure

getDSIDSU :: [DSUse]

DSUses of the data structure

type FunctionCall = (FunctionName, [Maybe VariableName])

Function call - name of the function, relevant arguments

data DSFun

Data structure for analysis info of a function definition

Constructors

DSF

Fields

getDSFFun :: FunctionDeclaration

Analyzed function declaration

getDSFCalls :: [FunctionCall]

FunctionCalls from the analyzed function

getDSFDSI :: [DSInfo]

DSInfo about the variables inside of the function, at this stage are not yet ready to obtain the information

data DSUse

Data structure use case

Constructors

DSU

Fields

getDSUName :: OperationName

Operation used

isHeavilyUsed :: Bool

Is it heavily used

isUserDependent :: Bool

Is it dependent on some external input (user, network, random, signals, etc.)

data FunctionDeclaration

Type for function definitions

Constructors

FunDecl

Fields

getFunName :: FunctionName

Name of the function

getFunType :: CTypeSpec

Return type of the function

getFunArgs :: [(VariableName, CTypeSpec)]

Names and types of the function arguments

type TermAnalyzer a = State TermAnalyzerState a

State monad with TermAnalyzerState

type Output = [(VariableName, DSUse)]

Term analyzer basic output, variables and DSUses

data TermAnalyzerState

State of the analyzer

Constructors

AS

Fields

getStateFunction :: Maybe FunctionDeclaration

Current function being analyzed

getStateFunNames :: [FunctionName]

All the function names

getStateVarNames :: [VariableName]

All the variable names --TODO are they only relevant ones? is it needed?

getStateCalls :: [FunctionCall]

FunctionCalls gathered through the analysis

append :: TermAnalyzerState -> TermAnalyzerState -> TermAnalyzerState

printDSI :: (String -> IO ()) -> DSInfo -> IO ()

Pretty print single DSInfo

printDSIAdvice :: (String -> IO ()) -> DSInfo -> IO ()

printRecommendationFromAnalysis :: (String -> IO ()) -> [DSInfo] -> IO ()

Pretty printer for the analyzer effects

printAdviceFromAnalysis :: (String -> IO ()) -> [DSInfo] -> IO ()

stupidMerge :: [DSInfo] -> [DSInfo]

Stupid merging of dsis --TODO remove this function, rewrite analyzeFunctions correctly

analyzeFunctions :: [DSFun] -> [DSInfo]

Merges the simple DSInfos based on function calls from the functions

lookupDSF :: [DSFun] -> FunctionName -> DSFun

Lookup DSFun by FunctionName

lookupDSI :: [DSInfo] -> VariableName -> FunctionName -> DSInfo

Lookup DSInfo by FunctionName and VariableName

bindFuncall :: [FunctionDeclaration] -> FunctionName -> [Maybe VariableName] -> [(VariableName, VariableName)]

Returns pairs of local variables bound to variables in a function that is called

generateDSI :: FunctionDeclaration -> [(VariableName, DSUse)] -> [DSInfo]

Generates simple DSInfos without the info from function calls

putVar :: VariableName -> TermAnalyzer ()

Start the state monad to create a DSFun for function generateDSF :: [FunctionName] -> FunctionDeclaration -> DSFun generateDSF = undefined generateDSF fnns fn = let (dsus, st) = runState (sumTerms step [getFunBody fn]) (AS fn fnns [] []) in DSF fn (getStateCalls st) (generateDSI fn dsus)

Function putting a variable definition in the context

getVar :: VariableName -> TermAnalyzerState -> Bool

Function returning True if the variable is already defined

parseMyFile :: FilePath -> IO CTranslUnit

Generate DSUses for a single Term step :: Term -> TermAnalyzer Output step = undefined

analyzeCTranslUnit :: CTranslUnit -> TermAnalyzer [Either Output DSFun]

analyzeCExtDecl :: CExtDecl -> TermAnalyzer (Either Output DSFun)

manalyzeCInit :: Maybe CInit -> TermAnalyzer Output

analyzeCInit :: CInit -> TermAnalyzer Output

analyzeCInitList :: CInitList -> TermAnalyzer Output

analyzeCTripleList :: [(Maybe CDeclr, Maybe CInit, Maybe CExpr)] -> TermAnalyzer Output

analyzeCFunDef :: CFunDef -> TermAnalyzer DSFun

analyzeCDerivedDeclarator :: CDerivedDeclr -> TermAnalyzer Output

analyzeCAttribute :: CAttr -> TermAnalyzer Output

analyzeCTypeQualifier :: CTypeQual -> TermAnalyzer Output

analyzeCArraySize :: CArrSize -> TermAnalyzer Output

manalyzeCDeclr :: Maybe CDeclr -> TermAnalyzer Output

analyzeCDeclr :: CDeclr -> TermAnalyzer Output

manalyzeCStatement :: Maybe CStat -> TermAnalyzer Output

analyzeCStatement :: CStat -> TermAnalyzer Output

analyzeCCompoundBlockItem :: CBlockItem -> TermAnalyzer Output

analyzeCConst :: CConst -> TermAnalyzer Output

manalyzeCExpr :: Maybe CExpr -> TermAnalyzer Output

isDsinfFunction :: FunctionName -> Bool

putCall :: FunctionName -> [CExpr] -> TermAnalyzer ()

Puts a function call into the state

analyzeCExpr :: CExpr -> TermAnalyzer Output

analyzeCBuiltin :: CBuiltin -> TermAnalyzer Output

analyzeCDesignator :: CDesignator -> TermAnalyzer Output

fmcs :: (Functor m, Monad m) => [m [a]] -> m [a]

Shortcut for fmap concat . sequence, useful in combining analysis of subterms