Defs.Common

Documentation

type Name = String

Type for names

type VariableName = String

Type for storing the variable names defined in a program

type FunctionName = String

Type for storing the function names defined in a program

startingFunction :: FunctionName

Name of the starting function