Defs.Structures

Documentation

data Structure

Data structure for keeping data structures

Constructors

DS

Fields

getDSName :: String

name of the data structure

getDSOps :: [DSOperation]

operations along with their complexities

data OperationName

Type for operation names

Constructors

InsertVal
Insert an element

DeleteByRef
Delete the element

FindByVal
Find the element by value

UpdateByRef
Update the value

DeleteExtremalVal
Delete the extreme value

ExtremalVal
Maximum or minimum

BoundByRef
Precedessor or successor

DecreaseValByRef
Update that decreases the value

Union
Union

Intersection
Intersection

Difference
Difference

SymDifference
Symmetric difference

Map
Map elements

Size
Checking the size

Empty
Checking the empiness

data ComplexityCharacteristics

Additional complexity qualifiers

Constructors

N
Normal time

A
Amortized time

E
Expected time

AE
Amortized expected time

type Complexity = (AsymptoticalComplexity, ComplexityCharacteristics)

Full complexity type

data AsymptoticalComplexity

Asymptotical complexity type, remembered as the exponent of the n and the number of stacked logarithms

Constructors

LinLog

Fields

getLin :: Integer

exponent of n

getLog :: Integer

number of the stacked logarithms

logs :: Integer -> String

Function to pretty print stacked logarithms

data DSOperation

Type for operation and its complexity

Constructors

Op

Fields

getOpName :: OperationName

Operation name

getComplexity :: Complexity

Complexity of the operation

compareDS :: Structure -> Structure -> [OperationName] -> Ordering

Function to check which of the two data structures, on given operations, is better