Defs.Util

Documentation

greenColor :: IO ()

Changes the color of the terminal output to green

blueColor :: IO ()

Changes the color of the terminal output to blue

yellowColor :: IO ()

Changes the color of the terminal output to yellow

redColor :: IO ()

Changes the color of the terminal output to red

cyanColor :: IO ()

Changes the color of the terminal output to cyan

resetColor :: IO ()

Resets the color of the terminal output

countElem :: Eq a => a -> [a] -> Integer

Count the number of occurences of a given element in the list

sequencesOfLen :: Eq a => [a] -> Integer -> [[a]]

Find all subsequences of length n and longer than sequence xs

split :: Eq a => [a] -> [a] -> [[a]]

A function inspired by python's string.split(). A list is split on a separator which is itself a list (not a single element).

maybeZipWith :: (a -> b -> Maybe c) -> [a] -> [b] -> [c]

Like zipWith only returns only those elements of type c that were qualified with Just