Recommend

Description

Module for recommending a data structure based on operations used

Documentation

recommendDS :: [OperationName] -> IO Structure

Recommends a data structure which is best for given operations opns. If there's more than one optimal structure, it chooses one at random

recommendAllDs :: [OperationName] -> [Structure]

Recommends all optimal data structures for given operations opns