Typechecker

Documentation

data TypeState

Variables with types, return value for a function, other declared functions

Constructors

TS

Fields

getStateReturn :: Maybe Type

getStateFunctions :: [Function]

getStateVariables :: [(VariableName, Type)]

type Typechecker a = State TypeState a

State monad to remember the TypeState

typecheckB :: [Term] -> Typechecker [Maybe Type]

Typecheck a term block

assertType :: Term -> Type -> Typechecker ()

Assert a type to a term, raise an error otherwise

typecheckF :: [Function] -> Function -> [Maybe Type]

Typecheck a function definition

typecheckFC :: FunctionName -> [Term] -> Typechecker (Maybe Type)

Typecheck a function call

typecheckR :: (String, Term) -> Typechecker (String, Type)

Typecheck one field of the record

typecheckT :: Term -> Typechecker (Maybe Type)

Typecheck a term, Nothing symbolizes the expressions without a value

relOp :: Term -> Term -> Typechecker (Maybe Type)

mathOp :: Term -> Term -> Typechecker (Maybe Type)

logOp :: Term -> Term -> Typechecker (Maybe Type)