Advice

Description

Module providing helpful tips about minimizing the complexity of the chosen data structure

Documentation

adviceDS :: [OperationName] -> [Advice]

Returns advice data for operations opns

printAdvice :: [OperationName] -> IO ()

Pretty prints effects of adviceDS

adviceDS' :: Integer -> [OperationName] -> [Advice]

Returns advice data for operations opns and at most n operations to forget about

printAdvice' :: Integer -> [OperationName] -> IO ()

Pretty prints effects of adviceDS'