Analyzer

Documentation

data DSInfo

Data structure for analysis info

Constructors

DSI

Fields

getDSINames :: [(FunctionName, VariableName)]

Variable holding the data structure --FIXME pointer copying

getDSIDSU :: [DSUse]

Data structure use cases

data DSFun

Data structure for function info

Constructors

DSF

Fields

getDSFFun :: Function

getDSFCalls :: [(FunctionName, [Maybe VariableName])]

getDSFDSI :: [DSInfo]

data DSUse

Data structure for use case info

Constructors

DSU

Fields

getDSUName :: OperationName

Operation used

isHeavilyUsed :: Bool

Is it heavily used

isUserDependent :: Bool

Is it dependent on some external input (user, network, random, signals, etc.)

type TermAnalyzer a = State TermAnalyzerState a

State monad with TermAnalyzerState

type TermAnalyzerOutput = [(VariableName, DSUse)]

data TermAnalyzerState

State of the analyzer

Constructors

AS

Fields

getStateFunction :: Function

Current function being analyzed

getStateFunNames :: [FunctionName]

All the other function names

getStateVarNames :: [VariableName]

All the variable names

getStateCalls :: [(FunctionName, [Maybe VariableName])]

Function calls gathered through the analysis

append :: TermAnalyzerState -> TermAnalyzerState -> TermAnalyzerState

printDSI :: (String -> IO ()) -> DSInfo -> IO ()

Pretty print single DSInfo

printDSIAdvice :: (String -> IO ()) -> DSInfo -> IO ()

printRecommendationFromAnalysis :: (String -> IO ()) -> [DSInfo] -> IO ()

Pretty printer for the analyzer effects

printAdviceFromAnalysis :: (String -> IO ()) -> [DSInfo] -> IO ()

stupidMerge :: [DSInfo] -> [DSInfo]

Stupid merging of dsis --TODO remove this function, rewrite analyzeFunctions correctly

analyze :: [Function] -> [DSInfo]

Runs everything that is needed to analyze a program

analyzeFunctions :: [DSFun] -> [DSInfo]

Merges the simple DSInfos based on function calls from the functions

lookupDSF :: [DSFun] -> FunctionName -> DSFun

Lookup DSFun by FunctionName

lookupDSI :: [DSInfo] -> VariableName -> FunctionName -> DSInfo

Lookup DSInfo by FunctionName and VariableName

bindFuncall :: [Function] -> FunctionName -> [Maybe VariableName] -> [(VariableName, VariableName)]

Returns pairs of local variables bound to variables in a function that is called

generateDSI :: Function -> [(VariableName, DSUse)] -> [DSInfo]

Generates simple DSInfos without the info from function calls

generateDSF :: [FunctionName] -> Function -> DSFun

Start the state monad to create a DSFun for function

stepBlock :: [Term] -> TermAnalyzer TermAnalyzerOutput

Analyze a block of terms using the state monad

sumTerms :: (Term -> TermAnalyzer TermAnalyzerOutput) -> [Term] -> TermAnalyzer TermAnalyzerOutput

Sums the DSUses using the step function and concating the results

putVar :: VariableName -> TermAnalyzer ()

Function putting a variable definition in the context

getVar :: VariableName -> TermAnalyzerState -> Bool

Function returning True if the variable is already defined

putCall :: FunctionName -> [Term] -> TermAnalyzer ()

Function putting a function call in the state

step :: Term -> TermAnalyzer TermAnalyzerOutput

Generate DSUses for a single Term