Defs.Common

Documentation

type Name = String

Type for names

newtype VariableName

Type for storing the variable names defined in a program

Constructors

V

Fields

unV :: String

newtype FunctionName

Type for storing the function names defined in a program

Constructors

F

Fields

unF :: String